Move duplicated set_up_repo integration-test helpers into git-tidy-core testutil
The set_up_repo helper was copy-pasted byte-identically across eight integration-test files (lfs / remote / stash / tag, each in integration_scan.rs and integration_clean.rs), and set_up_repo_with_remote was duplicated across git-branch-tidy's scan and clean tests. Lift both into git_tidy_core::testutil as pub fns (already gated behind the testutil feature, alongside TestRepo and git), and import them through each crate's common module. git-repo-tidy keeps its own set_up_repo(dir, name) helper, which has a different signature and is not a duplicate.
