+++ crates/git-tidy-core/src/testutil.rs
@@ -1560,6 +1560,68 @@ pub fn git(dir: &Path, args: &[&str]) -> String {
     String::from_utf8_lossy(&output.stdout).trim().to_string()
 }
 
+/// Set up a repo inside a scan directory so `discover_repos` finds it.
+///
+/// Returns `(tempdir, repo_path)`; the tempdir guards the lifetime of the
+/// on-disk repo and must be kept alive for the duration of the test.
+pub fn set_up_repo() -> (tempfile::TempDir, PathBuf) {
+    let dir = tempfile::tempdir().unwrap();
+    let base = dir.path().canonicalize().unwrap();
+
+    let scan_dir = base.join("projects");
+    std::fs::create_dir_all(&scan_dir).unwrap();
+
+    let repo = scan_dir.join("my-repo");
+    std::fs::create_dir_all(&repo).unwrap();
+    git(&repo, &["init", "-b", "main"]);
+    git(&repo, &["config", "user.email", "test@test.com"]);
+    git(&repo, &["config", "user.name", "Test"]);
+
+    std::fs::write(repo.join("README.md"), "# Test\n").unwrap();
+    git(&repo, &["add", "README.md"]);
+    git(&repo, &["commit", "-m", "Initial commit"]);
+
+    (dir, repo)
+}
+
+/// Set up a repo with an `origin` remote so default-branch detection works.
+///
+/// Returns `(tempdir, repo_path)`; the tempdir guards the lifetime of the
+/// on-disk repo and must be kept alive for the duration of the test.
+pub fn set_up_repo_with_remote() -> (tempfile::TempDir, PathBuf) {
+    let dir = tempfile::tempdir().unwrap();
+    let base = dir.path().canonicalize().unwrap();
+
+    // Create a bare "remote" repo
+    let bare = base.join("remote.git");
+    std::fs::create_dir_all(&bare).unwrap();
+    git(&bare, &["init", "--bare"]);
+
+    // Create the scan directory that holds the working repo
+    let scan_dir = base.join("projects");
+    std::fs::create_dir_all(&scan_dir).unwrap();
+
+    // Create the working repo inside scan_dir
+    let repo = scan_dir.join("my-repo");
+    std::fs::create_dir_all(&repo).unwrap();
+    git(&repo, &["init", "-b", "main"]);
+    git(&repo, &["config", "user.email", "test@test.com"]);
+    git(&repo, &["config", "user.name", "Test"]);
+
+    // Add remote
+    git(&repo, &["remote", "add", "origin", &bare.to_string_lossy()]);
+
+    // Initial commit
+    std::fs::write(repo.join("README.md"), "# Test\n").unwrap();
+    git(&repo, &["add", "README.md"]);
+    git(&repo, &["commit", "-m", "Initial commit"]);
+
+    // Push main to remote so origin/main exists
+    git(&repo, &["push", "-u", "origin", "main"]);
+
+    (dir, repo)
+}
+
 #[cfg(test)]
 mod mock_fidelity_tests {
     use super::*;
